requirements.txt
couchdb==1.2
python-dotenv==1.0.0
black==23.11.0
requests==2.31.0
inquirerpy==0.3.4
pandas==2.1.4
PyYAML==6.0
